Scripts : Stopped using var. Instead use let. Added error msg handling - might need to be a bit less naive. Added logIndex to processedLogs - useful when inserting logs that didnt match any transactions.
View :
Added tipsy.
Added popup to edge labels.
Shows an error msg if one is present and turns nodes and edges red that are in relation to that transaction (to from),
If no error message is present, there is currently just a placeholder popup.
Cleaned up graphDrawer a bit - could prolly use some more tho.